Attest: Add refs flag, improve help and command
This commit adds a new --refs flag to the attest subcommand. This lets
the user define an image reference to attach the attestation to. The help
text is also improved.

This commit also adds the --file reusable flag to allow attestations to be
written to a file.

Finally, this commit reshuffles the attest subcommand to use the pattern
of reusable flags in use in other subcommmands.

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>