Sign: Upload to tlog and capture sig data
This is a large commit that refactors the Sign() method of the attestation.
The main goal is to add two missing features:

1. Register the signature data to Rekor

After signing, we now register the signature in the sigstore transparency
log. This is essentail to allow for keyless verification.

2. New SignatureData Field

The attestation now has a new SignatureData field that captures the results
of the signing operation. This is required to make data like the cert and the
proof of inlclusion available externally (eg to record them in oci annotations).

The attestation.Sign() method has been heavily refactored but should be simppler
as the work it does is now broken into three internal functions:

 initSigning: creates context and options
 signAttestation: Performs the actual signing
 appendSignatureDataToTLog: Uploads data to rekor

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>